frontend-gui/requirements.txt
# packages in environment at //anaconda3/envs/660:
#
# Name                    Version                   Build  Channel
apscheduler               3.6.3                    py37_0  
asn1crypto                1.3.0                    py37_0  
backports                 1.0                        py_2  
backports.lzma            0.0.14           py37hd0afa68_0  
ca-certificates           2020.1.1                      0  
certifi                   2019.11.28               py37_1  
cffi                      1.14.0           py37hb5b8e2f_0  
chardet                   3.0.4                 py37_1003  
cryptography              2.8              py37ha12b0ac_0  
cycler                    0.10.0                     py_2    conda-forge
expat                     2.2.6                h0a44026_0  
freetype                  2.10.0               h24853df_1    conda-forge
gettext                   0.19.8.1             h15daf44_3  
html5lib                  1.0.1                    py37_0  
idna                      2.9                        py_1  
isodate                   0.6.0                      py_1  
jpeg                      9b                   he5867d9_2  
keepalive                 0.5                        py_1    conda-forge
kiwisolver                1.1.0            py37h770b8ee_0    conda-forge
libblas                   3.8.0               16_openblas    conda-forge
libcblas                  3.8.0               16_openblas    conda-forge
libcxx                    4.0.1                hcfea43d_1  
libcxxabi                 4.0.1                hcfea43d_1  
libedit                   3.1.20181209         hb402a30_0  
libffi                    3.2.1                h475c297_4  
libgfortran               4.0.0                         2    conda-forge
libiconv                  1.15                 hdd342a3_7  
liblapack                 3.8.0               16_openblas    conda-forge
libopenblas               0.3.9                h3d69b6c_0    conda-forge
libpng                    1.6.37               ha441bb4_0  
llvm-openmp               9.0.1                h28b9765_2    conda-forge
matplotlib                3.1.3                    py37_0  
matplotlib-base           3.1.3            py37h9aa3819_0  
ncurses                   6.2                  h0a44026_0  
numpy                     1.18.1           py37hde6bac1_0    conda-forge
openssl                   1.1.1f               h1de35cc_0  
pip                       20.0.2                   py37_1  
pycparser                 2.20                       py_0  
pyopenssl                 19.1.0                   py37_0  
pyparsing                 2.4.6                      py_0    conda-forge
pypubsub                  4.0.3                      py_0    conda-forge
pysocks                   1.7.1                    py37_0  
python                    3.7.6                h359304d_2  
python-dateutil           2.8.1                      py_0    conda-forge
python.app                2                       py37_10  
python_abi                3.7                     1_cp37m    conda-forge
pytz                      2019.3                     py_0  
rdflib                    4.2.2                 py37_1000    conda-forge
readline                  7.0                  h1de35cc_5  
requests                  2.23.0                   py37_0  
setuptools                46.0.0                   py37_0  
six                       1.14.0                   py37_0  
sparqlwrapper             1.8.2                 py37_1000    conda-forge
sqlite                    3.31.1               ha441bb4_0  
tk                        8.6.8                ha441bb4_0  
tornado                   6.0.4            py37h9bfed18_1    conda-forge
tzlocal                   2.0.0                    py37_0  
urllib3                   1.25.8                   py37_0  
webencodings              0.5.1                    py37_1  
wheel                     0.34.2                   py37_0  
wxpython                  4.0.4            py37hebc47a7_0  
xz                        5.2.4                h1de35cc_4  
zlib                      1.2.11               h1de35cc_3  
